requirements.txt
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0
face-recognition==1.3.0
dlib==19.24.2
pytesseract==0.3.10
easyocr==1.7.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
diffusers==0.24.0
accelerate==0.25.0
flask==3.0.0
flask-cors==4.0.0
matplotlib==3.8.2
scipy==1.11.4
imageio==2.31.6
jinja2==3.1.2


